refactor: Moved inbox-specific functions to Inbox handler
Refactored activity handler to support both inbox and outbox handlers. All inbox-specific functions were moved to a new Inbox handler. An Outbox handler will be added in a subsequent commit.

closes #183
